이메일 인증 삭제, 비밀번호 인증 완

+++ src/main/java/com/gobang/gobang/domain/personal/dto/response/StatsResponse.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class StatsResponse {
-    private int totalPoints;
-    private int totalReviews;
+    private Long totalPoints;
+    private Long totalReviews;
     private String membershipLevel;
 }
